Got things working
Got things working by using TeensySmartMatrix instead of generic 'Serial' driver from the README.md.  Updated test.py (didn't see that when working from code in the README) to be runnable with new BP and updated sample code in README.md to be runnable, but now some of the lower comments are out of date.